requirements.txt
Django~=4.1.6
Pillow
djangorestframework~=3.14.0